+++ src/routes/MemberProfile/MemberProfile.connector.js
@@ -1,5 +1,6 @@
 import { connect } from 'react-redux'
 import { push } from 'connected-react-router'
+import { postUrl } from 'util/navigation'
 import blockUser from 'store/actions/blockUser'
 import isPendingFor from 'store/selectors/isPendingFor'
 import getPreviousLocation from 'store/selectors/getPreviousLocation'
@@ -10,6 +11,8 @@ import {
   FETCH_MEMBER_POSTS,
   FETCH_MEMBER_COMMENTS,
   FETCH_MEMBER_VOTES,
+  FETCH_PROJECTS,
+  fetchProjects,
   getPresentedPerson
 } from './MemberProfile.store'
 
@@ -25,7 +28,8 @@ export function mapStateToProps (state, props) {
     FETCH_RECENT_ACTIVITY,
     FETCH_MEMBER_POSTS,
     FETCH_MEMBER_COMMENTS,
-    FETCH_MEMBER_VOTES
+    FETCH_MEMBER_VOTES,
+    FETCH_PROJECTS
   ], state)
   const personLoading = isPendingFor(fetchPerson, state)
 
@@ -35,15 +39,22 @@ export function mapStateToProps (state, props) {
     personLoading,
     contentLoading,
     person,
+    projects: state.MemberProfile.projects,
     currentUser: getMe(state),
     previousLocation: getPreviousLocation(state)
   }
 }
 
-export const mapDispatchToProps = {
-  fetchPerson,
-  blockUser,
-  push
+export function mapDispatchToProps (dispatch, props) {
+  const { routeParams, querystringParams } = props
+
+  return {
+    fetchPerson: (id) => dispatch(fetchPerson(id)),
+    fetchProjects: (context, contextId, viewerId) => dispatch(fetchProjects(context, contextId, viewerId)),
+    blockUser: (id) => dispatch(blockUser(id)),
+    push,
+    showDetails: (id) => dispatch(push(postUrl(id, routeParams, querystringParams)))
+  }
 }
 
 export function mergeProps (stateProps, dispatchProps, ownProps) {